[UPD] utils refactor
