+++ utils/perms.py
@@ -2,6 +2,7 @@ import nextcord
 import nextcord.ext.commands
 
 from utils.config import Configuration
+from utils.database import db_session, StaffBlacklist
 from utils.sersi_embed import SersiEmbed
 
 config = Configuration.from_yaml_file("./persistent_data/config.yaml")
@@ -232,3 +233,16 @@ async def cb_is_dark_mod(interaction) -> bool:
 
 async def cb_is_cet(interaction) -> bool:
     return await permcheck(interaction, is_cet)
+
+
+def blacklist_check(user: nextcord.Member):
+    with db_session() as session:
+        blacklisted = (
+            session.query(StaffBlacklist).filter_by(blacklisted_user=user.id).first()
+        )
+
+        if blacklisted:
+            return True
+
+        else:
+            return False
